gpd renamed to semantic segmentation detector
