+++ ROAR_simulation/roar_autonomous_system/perception_module/semantic_segmentation_detector.py
