+++ ROAR_simulation/roar_autonomous_system/planning_module/local_planner/gpd_only_local_planner.py
@@ -5,7 +5,7 @@ from ROAR_simulation.roar_autonomous_system.utilities_module.vehicle_models impo
     Vehicle,
 )
 from ROAR_simulation.roar_autonomous_system.control_module.controller import Controller
-from ROAR_simulation.roar_autonomous_system.perception_module.ground_plane_detector import (
+from ROAR_simulation.roar_autonomous_system.perception_module.semantic_segmentation_detector import (
     SemanticSegmentationDetector,
 )
 from ROAR_simulation.roar_autonomous_system.utilities_module.vehicle_models import (
